pyproject.toml
[tool.poetry]
name = "logseq_chat"
version = "0.1.0"
description = "Chat with an LLM about your Logseq notes"
authors = ["Bradley Heilbrun <108751498+bheilbrun@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.scripts]
logseq_chat = "logseq_chat.main:main"

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
faiss-cpu = "^1.8.0.post1"
langchain = "^0.2.11"
langchain-anthropic = "^0.1.22"
langchain-community = "^0.2.10"
langchain-core = "^0.2.28"
langchain-openai = "^0.1.20"
langchain-text-splitters = "^0.2.2"
python-dotenv = "^1.0.1"
watchdog = "^4.0.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.5"
ipykernel = "^6.29.5"
deptry = "^0.18.0"
mypy = "^1.11.1"
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
